+++ src/Mscc.GenerativeAI/Types/Generative/Part.cs
@@ -1,5 +1,4 @@
 ﻿using System.Diagnostics;
-using System.Text.Json.Serialization;
 
 namespace Mscc.GenerativeAI
 {
@@ -92,5 +91,16 @@ namespace Mscc.GenerativeAI
         {
             return new FileData { FileUri = uri, MimeType = mimetype };
         }
+    }
+
+    /// <summary>
+    /// A datatype containing data that is part of a multi-part `TuningContent` message.
+    /// This is a subset of the Part used for model inference, with limited type support.
+    /// A `Part` consists of data which has an associated datatype.
+    /// A `Part` can only contain one of the accepted types in `Part.data`.
+    /// </summary>
+    [DebuggerDisplay("{Text}")]
+    public class TuningPart : Part
+    {
     }
 }