+++ src/Mscc.GenerativeAI/Types/Generative/Schema.cs
@@ -15,9 +15,9 @@ namespace Mscc.GenerativeAI
         /// Optional. The format of the data.
         /// This is used only for primitive datatypes.
         /// Supported formats:
-        ///  for NUMBER type: float, double
-        ///  for INTEGER type: int32, int64
-        ///  for STRING type: enum, date-time
+        /// for NUMBER type: float, double
+        /// for INTEGER type: int32, int64
+        /// for STRING type: enum, date-time
         /// </summary>
         public string Format { get; set; } = "";
         /// <summary>