+++ tests/Mscc.GenerativeAI/GoogleAi_GeminiPro_Should.cs
@@ -23,7 +23,7 @@ namespace Test.Mscc.GenerativeAI
     [Collection(nameof(ConfigurationFixture))]
     public class GoogleAiGeminiProShould(ITestOutputHelper output, ConfigurationFixture fixture)
     {
-        private readonly string _model = Model.Gemini15Pro;
+        private readonly string _model = Model.Gemini20FlashExperimental;
 
         [Fact]
         public void Initialize_GoogleAI()
@@ -158,7 +158,6 @@ namespace Test.Mscc.GenerativeAI
         }
 
         [Theory]
-        [InlineData(Model.Gemini10Pro001)]
         [InlineData(Model.GeminiProVision)]
         [InlineData(Model.BisonText)]
         [InlineData(Model.BisonChat)]
@@ -192,7 +191,6 @@ namespace Test.Mscc.GenerativeAI
         }
 
         [Theory]
-        [InlineData(Model.Gemini10Pro001)]
         [InlineData(Model.GeminiProVision)]
         [InlineData(Model.BisonText)]
         [InlineData(Model.BisonChat)]
@@ -258,7 +256,7 @@ namespace Test.Mscc.GenerativeAI
             // Arrange
             var prompt = "Tell me 4 things about Taipei. Be short.";
             var googleAi = new GoogleAI(apiKey: "WRONG_API_KEY");
-            var model = googleAi.GenerativeModel(model: Model.Gemini10Pro001);
+            var model = googleAi.GenerativeModel(model: Model.GeminiPro);
             model.ApiKey = fixture.ApiKey;
 
             // Act
@@ -277,7 +275,7 @@ namespace Test.Mscc.GenerativeAI
             // Arrange
             var prompt = "Tell me 4 things about Taipei. Be short.";
             var googleAi = new GoogleAI(apiKey: "WRONG_API_KEY");
-            var model = googleAi.GenerativeModel(model: Model.Gemini10Pro001);
+            var model = googleAi.GenerativeModel(model: Model.GeminiPro);
             await Assert.ThrowsAsync<HttpRequestException>(() => model.GenerateContent(prompt));
             
             // Act
@@ -297,7 +295,7 @@ namespace Test.Mscc.GenerativeAI
             // Arrange
             var prompt = "Write a story about a magic backpack.";
             var googleAi = new GoogleAI(apiKey: fixture.ApiKey);
-            var model = googleAi.GenerativeModel(model: Model.Gemini10Pro001);
+            var model = googleAi.GenerativeModel(model: Model.GeminiPro);
             var generationConfig = new GenerationConfig() { MaxOutputTokens = 20 };
 
             // Act
@@ -307,7 +305,26 @@ namespace Test.Mscc.GenerativeAI
             response.Should().NotBeNull();
             response.Candidates.Should().NotBeNull().And.HaveCount(1);
             response.Text.Should().BeEmpty();
-            // output.WriteLine(response?.Text);
+            output.WriteLine(response?.Text);
+        }
+
+        [Fact]
+        public async Task Generate_Content_Logprobs()
+        {
+            // Arrange
+            var prompt = "Write a story about a magic backpack.";
+            var googleAi = new GoogleAI(apiKey: fixture.ApiKey);
+            var model = googleAi.GenerativeModel(model: Model.Gemini15Pro002);
+            var generationConfig = new GenerationConfig() { ResponseLogprobs = true };
+
+            // Act
+            var response = await model.GenerateContent(prompt, generationConfig);
+
+            // Assert
+            response.Should().NotBeNull();
+            response.Candidates.Should().NotBeNull().And.HaveCount(1);
+            response.Text.Should().BeEmpty();
+            output.WriteLine(response?.Text);
         }
 
         [Fact]
@@ -316,7 +333,7 @@ namespace Test.Mscc.GenerativeAI
             // Arrange
             var prompt = "Write a story about a magic backpack.";
             var googleAi = new GoogleAI(apiKey: fixture.ApiKey);
-            var model = googleAi.GenerativeModel(model: Model.Gemini10Pro001);
+            var model = googleAi.GenerativeModel(model: Model.GeminiPro);
             var generationConfig = new GenerationConfig() { MaxOutputTokens = 20 };
 
             // Act
@@ -843,6 +860,49 @@ namespace Test.Mscc.GenerativeAI
             output.WriteLine(string.Join(Environment.NewLine,
                 response.Candidates[0].Content.Parts.Select(x => 
                     x.Text)
+//                    .Where(t => !string.IsNullOrEmpty(t))
+                    .ToArray()));
+        }
+
+        [Theory]
+        [InlineData(Model.Gemini15Flash)]
+        [InlineData(Model.Gemini20Flash)]
+        [InlineData(Model.Gemini20FlashLite)]
+        [InlineData(Model.Gemini20FlashThinking)]
+        [InlineData(Model.Gemini20Pro)]
+        // Ref: https://ai.google.dev/api/generate-content#code-execution
+        public async Task Generate_Content_Code_Execution_using_FileAPI(string modelName)
+        {
+            // Arrange
+            var prompt = "Calculate the sum of the column 'effect'.";
+            var genAi = new GoogleAI(fixture.ApiKey);
+            var model = genAi.GenerativeModel(modelName);
+            var filePath = Path.Combine(Environment.CurrentDirectory, "payload", "sampledata.csv");
+            var file = await ((GoogleAI)genAi).UploadFile(filePath);
+            output.WriteLine($"File: {file.File.Name}\tName: '{file.File.DisplayName}'");
+            var request = new GenerateContentRequest(prompt)
+            {
+                GenerationConfig = new GenerationConfig()
+                {
+                    Temperature = 1f,
+                    TopK = 40,
+                    TopP = 0.95f,
+                    MaxOutputTokens = 8192,
+                    ResponseMimeType = "text/plain"
+                },
+                Tools = [new Tool { CodeExecution = new() }]
+            };
+            request.AddMedia(file.File);
+
+            // Act
+            var response = await model.GenerateContent(request);
+
+            // Assert
+            response.Should().NotBeNull();
+            response.Candidates.Should().NotBeNull().And.HaveCount(1);
+            output.WriteLine(string.Join(Environment.NewLine,
+                response.Candidates[0].Content.Parts.Select(x => 
+                        x.Text)
 //                    .Where(t => !string.IsNullOrEmpty(t))
                     .ToArray()));
         }
@@ -1542,11 +1602,11 @@ namespace Test.Mscc.GenerativeAI
         {
             // Arrange
             var googleAI = new GoogleAI(accessToken: fixture.AccessToken);
-            var model = googleAI.GenerativeModel(model: Model.Gemini10Pro001);
+            var model = googleAI.GenerativeModel(model: Model.GeminiPro);
             model.ProjectId = fixture.ProjectId;
             var request = new CreateTunedModelRequest()
             {
-                BaseModel = $"{Model.Gemini10Pro001.SanitizeModelName()}",
+                BaseModel = $"{Model.GeminiPro.SanitizeModelName()}",
                 DisplayName = "Autogenerated Test model",
                 TuningTask = new()
                 {
@@ -1590,7 +1650,7 @@ namespace Test.Mscc.GenerativeAI
         {
             // Arrange
             var googleAI = new GoogleAI(accessToken: fixture.AccessToken);
-            var model = googleAI.GenerativeModel(model: Model.Gemini10Pro001);
+            var model = googleAI.GenerativeModel(model: Model.GeminiPro);
             model.ProjectId = fixture.ProjectId;
             var parameters = new HyperParameters() { BatchSize = 2, LearningRate = 0.001f, EpochCount = 3 };
             var dataset = new List<TuningExample>
@@ -1607,7 +1667,7 @@ namespace Test.Mscc.GenerativeAI
                 new() { TextInput = "thirteen", Output = "fourteen" },
                 new() { TextInput = "seven", Output = "eight" },
             };
-            var request = new CreateTunedModelRequest(Model.Gemini10Pro001, 
+            var request = new CreateTunedModelRequest(Model.GeminiPro, 
                 "Simply autogenerated Test model",
                 dataset,
                 parameters);